Implementation files
Move logic from header files into new .cpp implementation files.
