+++ src/spv/data/data.hpp
@@ -9,13 +9,8 @@
 #include <cassert>
 #include <cstdint>
 #include <limits>
-#include <stdexcept>
 
 #include "../../util/spirv.hpp"
-#include "../../values/aggregate.hpp"
-#include "../../values/coop-matrix.hpp"
-#include "../../values/primitive.hpp"
-#include "../../values/string.hpp"
 #include "../../values/type.hpp"
 #include "../../values/value.hpp"
 
@@ -82,24 +77,7 @@ public:
         return name;
     }
 
-    void initValue(const Type& t) {
-        // Only initialize the value once
-        assert(this->val == nullptr);
-
-        // Construct the value from the given type
-        // For whatever reason, the SPIR-V spec says that the type of each OpVariable must be an OpTypePointer, although
-        // it is actually storing the value.
-        // Therefore, before we construct, we need to dereference the pointer
-        if (t.getBase() != DataType::POINTER)
-            throw std::invalid_argument("Cannot initialize variable with non-pointer type!");
-        this->val = t.getPointedTo().construct();
-        auto set_unsized = [&](Value& seen) {
-            if (seen.getType().getBase() == DataType::COOP_MATRIX)
-                static_cast<CoopMatrix&>(seen).setUnsized();
-            return true;
-        };
-        val->recursiveApply(set_unsized);
-    }
+    void initValue(const Type& t);
 
     bool isThreaded() const {
         return storage == spv::StorageClassPrivate || storage == spv::StorageClassFunction;
@@ -149,27 +127,7 @@ public:
         return UNSET;
     }
 
-    [[nodiscard]] Value* asValue() const override {
-        // Represent this variable with its value, storage class, and if set, name
-        // Don't currently display decorations although they could be helpful.
-        std::vector<Value*> elements;
-        std::vector<std::string> names;
-
-        if (!name.empty()) {
-            names.push_back("name");
-            elements.push_back(new String(name));
-        }
-
-        names.push_back("value");
-        Value* cloned = val->getType().construct();
-        cloned->copyFrom(*val);
-        elements.push_back(cloned);
-
-        names.push_back("storage-class");
-        elements.push_back(new Primitive(static_cast<int32_t>(storage)));
-
-        return new Struct(elements, names);
-    }
+    [[nodiscard]] Value* asValue() const override;
 };
 
 class Function : public Valuable {
@@ -189,27 +147,7 @@ public:
         return location;
     }
 
-    [[nodiscard]] Value* asValue() const override {
-        std::vector<Value*> elements;
-        std::vector<std::string> names;
-        // Populate the representative struct with three fields:
-        // - name (only use if has been set (ie, is not ""))
-        // - type
-        // - location
-
-        if (!name.empty()) {
-            names.push_back("name");
-            elements.push_back(new String(name));
-        }
-
-        names.push_back("types");
-        elements.push_back(type->asValue());
-
-        names.push_back("location");
-        elements.push_back(new Primitive(location));
-
-        return new Struct(elements, names);
-    }
+    [[nodiscard]] Value* asValue() const override;
 };
 
 struct EntryPoint final : public Function {
@@ -266,26 +204,8 @@ public:
 
     // Fetching of Values must be able to fetch spec constants, which are saved as program inputs but also need to be
     // usable like regular values.
-    Value* getValue() {
-        if (type == DType::VALUE)
-            return static_cast<Value*>(raw);
-        if (type == DType::VARIABLE) {
-            auto var = static_cast<Variable*>(raw);
-            if (var->isSpecConst())
-                return &var->getVal();
-        }
-        return nullptr;
-    }
-    const Value* getValue() const {
-        if (type == DType::VALUE)
-            return static_cast<const Value*>(raw);
-        if (type == DType::VARIABLE) {
-            auto var = static_cast<Variable*>(raw);
-            if (var->isSpecConst())
-                return &var->getVal();
-        }
-        return nullptr;
-    }
+    Value* getValue();
+    const Value* getValue() const;
 
     // Convenience function to not need to define the Data for each use
     template<typename T>
@@ -304,70 +224,9 @@ public:
         this->own = own;
     }
 
-    void clear() {
-        if (own) {
-            switch (type) {
-            default:
-                assert(false);
-                break;
-            case DType::UNDEFINED:
-                // do nothing since there is no data to delete
-                break;
-            case DType::VARIABLE:
-                delete static_cast<Variable*>(raw);
-                break;
-            case DType::FUNCTION:
-                delete static_cast<Function*>(raw);
-                break;
-            case DType::ENTRY:
-                delete static_cast<EntryPoint*>(raw);
-                break;
-            case DType::VALUE:
-                delete static_cast<Value*>(raw);
-                break;
-            case DType::TYPE:
-                delete static_cast<Type*>(raw);
-                break;
-            }
-        }
+    void clear();
 
-        type = DType::UNDEFINED;
-    }
-
-    Data clone() const {
-        // NOTE: If this data does not own its value, then cloning it will *not* grant ownership for the clone. This
-        // should be ok since we only use weak data for RT where we are referencing the main stage which we know will
-        // outlive any cloned substage.
-        if (!own && type != DType::UNDEFINED) {
-            Data other;
-            other.raw = this->raw;
-            other.own = false;
-            other.type = this->type;
-            return other;
-        }
-
-        switch (type) {
-        default:
-            assert(false);
-            return Data();
-        case DType::UNDEFINED:
-            return Data();
-        case DType::VARIABLE:
-            return Data(new Variable(*static_cast<Variable*>(raw)));
-        case DType::FUNCTION:
-            return Data(new Function(*static_cast<Function*>(raw)));
-        case DType::ENTRY:
-            return Data(new EntryPoint(*static_cast<EntryPoint*>(raw)));
-        case DType::VALUE: {
-            Value* before = static_cast<Value*>(raw);
-            Value* after = before->getType().construct();
-            after->copyFrom(*before);
-            return Data(after);
-        }
-        case DType::TYPE:
-            return Data(new Type(*static_cast<Type*>(raw)));
-        }
-    }
+    Data clone() const;
 
     /// @brief Move the data held by other into this, transferring ownership
     /// @param other transferred from. Cleared before return.
