+++ src/spv/frame.hpp
@@ -8,7 +8,6 @@
 
 #include <stdexcept>
 
-#include "../values/primitive.hpp"
 #include "../values/raytrace/accel-struct.hpp"
 #include "../values/value.hpp"
 #include "data/manager.hpp"
@@ -29,6 +28,7 @@ inline const char* to_string(RtStageKind kind) {
     case RtStageKind::CALLABLE:
         return "callable";
     default:
+        assert(false && "Invalid RtStageKind!");
         return "invalid";
     }
 }
@@ -87,35 +87,15 @@ public:
         , first(true) {}
     Frame(const Frame&) = delete;
     Frame& operator=(const Frame&) = delete;
-    ~Frame() {
-        if (view != nullptr) {
-            view->getSource()->destroyView(view);
-            view = nullptr;  // to prevent double deletion or after-deletion use
-        }
-        if (this->rt.data != nullptr) {
-            delete this->rt.data;
-            this->rt.data = nullptr;
-        }
-    }
+    ~Frame();
 
-    unsigned getPC() const {
+    inline unsigned getPC() const {
         return pc;
     }
 
-    Data& getArg() noexcept(false) {
-        if (argCount >= args.size())
-            throw std::runtime_error("No more args to use!");
-        ++pc;
-        return *args[argCount++];
-    }
+    Data& getArg() noexcept(false);
 
-    void incPC() noexcept(false) {
-        if (first)
-            first = false;
-        else if (argCount < args.size())
-            throw std::runtime_error("Unused function argument(s)!");
-        ++pc;
-    }
+    void incPC() noexcept(false);
 
     void setPC(unsigned pc) noexcept(false) {
         if (argCount < args.size())
@@ -151,41 +131,10 @@ public:
         return rt.trigger;
     }
 
-    void triggerRaytrace(RtStageKind stage, unsigned index, Value* payload, Value* hit_attrib, AccelStruct& as) {
-        this->rt.trigger = stage;
-        this->rt.index = index;
-        this->rt.as = &as;
-        this->rt.result = payload;
-        this->rt.hitAttribute = hit_attrib;
-        if (this->rt.data != nullptr) {
-            delete this->rt.data;
-            this->rt.data = nullptr;
-        }
-    }
-    void triggerCallable(unsigned index, Value* callable, AccelStruct* as) {
-        this->rt.trigger = RtStageKind::CALLABLE;
-        this->rt.index = index;
-        this->rt.as = as;
-        this->rt.result = callable;
+    void triggerRaytrace(RtStageKind stage, unsigned index, Value* payload, Value* hit_attrib, AccelStruct& as);
+    void triggerCallable(unsigned index, Value* callable, AccelStruct* as);
 
-        // hit attribute is never used by callable, so we reuse it to track whether this frame is entry or exit
-        Primitive dummy(0);
-        this->rt.hitAttribute = static_cast<Value*>(&dummy);
-
-        if (this->rt.data != nullptr) {
-            delete this->rt.data;
-            this->rt.data = nullptr;
-        }
-    }
-
-    void disableRaytrace() {
-        this->rt.trigger = RtStageKind::NONE;
-        this->rt.index = 0;
-        this->rt.as = nullptr;
-        this->rt.result = nullptr;
-        delete this->rt.data;
-        this->rt.data = nullptr;
-    }
+    void disableRaytrace();
 
     bool isCallableReturn() const {
         if (this->rt.trigger == RtStageKind::NONE)
